Fixed some file buffer flushing bugs.  However, the code now works and produces data comparable to Ogg Vorbis 1.0rc3.
There appear to be one or two small math problems lurking around somewhere, but
they dont seem to affect quality (+/- 1 in some of the PCM values...looks
like a rounding error).

Need to finish porting synchronized() to C#'s locking mechanism, and editing
the comments in the files.  Also, should find a way to clean up all the
internal methods/variables.  csogg is a huge abuser of this...

svn path=/trunk/csvorbis/; revision=4103