Workaround for mcs bug.  Basically constructs like:
a[i++]+=g;

would expand to:

a[i++] = a[i++] + g;

which lead to bad things happening.  So, occurences of that (1 in
CodeBook.cs, one in Lsp.c, and 4 in Drft.cs), were changed to the
generic:

a[i] += g;
i++;

This sounds like it works on all streams beta4 onward.

With this change, csvorbis compiles and works with mcs.

svn path=/trunk/csvorbis/; revision=4688